Pmatch: Remove bad abstractions in the form-building code #669693
Unforuntately, this puts back duplication, but there was already duplication
between pmatch stand-alone and combined, which this did not solve.

We should try to eliminate the duplication in future, but for now, at least
we have vanilla forms code.

As part of this change, I have made the display of fields consistent
between the various questions types. Pmatch, pmatchexamples and
pmatch-in-combined don't have exactly the same options, but the
options they do have are now displayed in the same order and in the same way.